Added the mostly working web UI.
A navigation bar with a tab for submitting an answer to the sample question exists. The results page for the question does not currently have a proper pie chart for displaying results, but the results are displayed on the page in text form.

Navigation bar also has a link to this repo.