Added the mostly working web UI, templates for each page, and a navigation bar.
The navigation bar has a tab for submitting an answer to the sample question exists. The results page for the question does not currently have a proper pie chart for displaying results, but the results are displayed on the page in text form.

Navigation bar also has a link to this repo.